Aplicando MVC em Presentation.java e MenuController.java
Aplicação de boas práticas na classe "Presentation" e "MenuController". Também foram feitas mudanças nas classes para respeitar a lei de responsabilidade única do S.O.L.I.D. Para isso foi criada uma nova classe "MenuView" para lidar com as responsabilidades extras que estavam presentes em MenuController.
