+++ src/main/java/br/upe/enenhariasoftware/psw/jabberpoint/view/SlideViewerFrame.java
@@ -48,7 +48,7 @@ public class SlideViewerFrame extends JFrame {
 
     getContentPane().add(slideViewerComponent);
     addKeyListener(new KeyController(presentation));
-    setMenuBar(new MenuController(this, presentation));
+    setMenuBar(new MenuView(new MenuController(this, presentation)));
     setSize(new Dimension(WIDTH, HEIGHT));
 
     setVisible(true);
